cheri: Make net/ compile.
Make net/ compile for CHERI, add casts etc.

Disable CONFIG_RDS for the time being. The UAPI issues related
to that need more in-depth analysis.

Add a hack to the xt_counters data structure. It is either used
as two u64 or as one pointer. Thus we can fit the pointer in there.

+++ fs/afs/rxrpc.c
@@ -712,7 +712,7 @@ void afs_charge_preallocation(struct work_struct *work)
 		if (rxrpc_kernel_charge_accept(net->socket,
 					       afs_wake_up_async_call,
 					       afs_rx_attach,
-					       (unsigned long)call,
+					       (uintptr_t)call,
 					       GFP_KERNEL,
 					       call->debug_id) < 0)
 			break;
